requirements.txt
dj-database-url==0.5.0
Django==2.2
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.9.2
mysql==0.0.2
mysqlclient==2.0.1
